chore: テスト用型定義を分離し Jest 設定を調整
